A Facet now has a __name__ that is managed by its facet manager.
The facet's name is unique within the collection of facets, and the name
will not be allocated to another facet if this one is removed and another
one added. Basically, they are monotonically increasing.

You can retrieve a particular facet by its name.

Facet's names are used in URLs for facets.