Change the database version back to 0.8 as databases from 0.8 are upgradable.
Was a bug in the rules for schoolbell 0.8 which is triggered buy this sequence:

apt-get install schoolbell -- (0.8)
apt-get remove schoolbell
apr-get install schoolbell -- (0.8 or 0.9)

At which point the user is told the database is incompatible and needs to be
removed.

`apt-get upgrade schoolbell` and schooltool are unaffected.

The only way of fixing I can think of it to not do the check at all. In which
case upgrades from 0.7 > 0.9 will break.

I hope nobody who cares runs into this.