added isToday() method to CalendarDay class for issue180
returns a boolean that you can use for setting the ID or class of a day in
week/month/year view.  Added it to week/month but had a problem with getting
it to show up in renderRow for year view.  Its not exposed in CSS, but if you
view source on week/month views you'll see id="today" on the correct day.