Optimize sample attendance data so it is not excessively painful, and enable it.  It still doubles the time taken to run functional tests on my machine.
The principal change is Ignas's idea to reuse PRESENT attendance records for
all section members.  I'm not entirely confident the optimization is always
safe; but it will definitely work if we record presence for all members of a
particular section in a single transaction, which is what we do now.