Factor out a lot out of schooltool.app.browser into schooltool.skin.
The goal is to eventually have a standalone schooltool.skin that can
be easily developed further without worrying too much about the rest of
schooltool. The goal is also to clean up schooltool.app.browser. Right
now what remains there is calendaring code, some skin setup code, 
and other bits and pieces.

The goal is also make schooltool.skin as standalone as possible.

We haven't reached the goals yet completely, but we're a good way and
the tests all work again, so checking in.