* generalized formlib based forms and moved them into schooltool.skin
* add a password setting widget, which makes it easier to use it without
  complicated form wizardy.

* use the password setting widget and the generalized form
  in schooltool.person in a password_edit.html view.

Note that we did not remove the edit.html view yet in schooltool.person, but
we're preparing to so by replacing the password setting functionality.
We've already replaced the fullname setting functionality in 
the demographics package. We still need to replace the photo upload
functionality but it doesn't work anyway. :)